+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next version
+# 0.12.0
 + Add scss
 + Save expire token date
 + Add check and refresh token